Lifting state up to the Game component
