+++ react-app/src/app/screens/Game/utils.js
